+++ Global Domination/InboxView.cpp
@@ -14,13 +14,13 @@ namespace global_domination {
 
 	void InboxView::initialize()
 	{
-		SDL_Rect list_box_client_area{ client_area_.w * 0.05, client_area_.h * 0.15, client_area_.w * 0.25, client_area_.h * 0.8 };
+		SDL_Rect list_box_client_area{ (int)(client_area_.w * 0.05), (int)(client_area_.h * 0.15), (int)(client_area_.w * 0.25), (int)(client_area_.h * 0.8) };
 		std::shared_ptr<SDLListBox<Message>> message_list = std::make_shared<SDLListBox<Message>>(parent_, list_box_client_area);
 		message_list->setId(INBOX_LIST);
 		addChildControl(message_list);
 
-		addLabel("", client_area_.w * 0.6, client_area_.h * 0.3, SELECTED_MESSAGE_TITLE_LABEL);
-		addLabel("", client_area_.w * 0.6, client_area_.h * 0.4, SELECTED_MESSAGE_TEXT_LABEL, 12);
+		addLabel("", (int)(client_area_.w * 0.6), (int)(client_area_.h * 0.3), SELECTED_MESSAGE_TITLE_LABEL);
+		addLabel("", (int)(client_area_.w * 0.6), (int)(client_area_.h * 0.4), SELECTED_MESSAGE_TEXT_LABEL, 12);
 
 		updateMessageList();
