+++ Global Domination/SDLDropDownMenu.h
@@ -117,7 +117,7 @@ namespace global_domination
 		{
 			// make menu width slightly longer than the estimated text length
 			int text_length_pixels = 140;
-			SDL_Rect menu_client_area{ kMenuItemPosX, kMenuItemPosY, text_length_pixels * 1.2, kMenuItemHeight * menu_items_.size() };
+			SDL_Rect menu_client_area{ kMenuItemPosX, kMenuItemPosY, (int)(text_length_pixels * 1.2), kMenuItemHeight * (int)(menu_items_.size()) };
 
 			SDL_SetRenderDrawColor(renderer, 255, 255, 255, 0xFF);
 			SDL_RenderFillRect(renderer, &menu_client_area);
